Add some rust guests
