+++ README.md
@@ -14,26 +14,27 @@ Invoke: `curl -X GET -i http://127.0.0.1:8000/python-svc/secret`
 # Writing guests for RVM
 You can write guests in any language you want, as long as it compiles to webassembly.
 This example uses python.
+There are other rust examples in the repo that are more straight forward, those can simply be compiled with something like `cargo build -p http-rust --target wasm32-wasip2`.
 
 ### 1. Implement
 
 Write your app and import the things you need.
-There's an example in `guests/http_server.py` that implements a HTTP server that can be run in RVM.
+There's an example in `guests/http-python.py` that implements a HTTP server that can be run in RVM.
 Right now, the RVM expects all guests to be a HTTP proxy.
 Every time it receives an `invoke` request it will run `IncomingHandler::handle` in your guest, with a forwarded HTTP request.
 
 ### 2. Build
 1. Make sure you have `componentize-py`, which can be installed via `pip install componentize-py`
-2. Build the wasm module `cd guests && componentize-py -d ../wit -w rvm componentize http_server -o python-svc.wasm`
+2. Build the wasm module `cd guests && componentize-py -d ../wit -w rvm componentize http-python -o http-python.wasm`
 
 ### 3. Deploy
-`curl --data-binary "@python-svc.wasm" localhost:8002/deploy/python-svc` 
+`curl --data-binary "@http-python.wasm" localhost:8002/deploy/http-python` 
 
 ### 4. Talk to your deployed app
 
-* Get the SHA of some page: - `curl -X GET -i -H 'url: https://webassembly.github.io/spec/core/' http://127.0.0.1:8000/python-svc/hash-all`
-* Echo back a body - `curl -X POST -i http://127.0.0.1:8000/python-svc/echo -d "xd"`
-* Print a secret provided by the host - `curl -X GET -i http://127.0.0.1:8000/python-svc/secret`
+* Get the SHA of some page: - `curl -X GET -i -H 'url: https://webassembly.github.io/spec/core/' http://127.0.0.1:8000/http-python/hash-all`
+* Echo back a body - `curl -X POST -i http://127.0.0.1:8000/http-python/echo -d "xd"`
+* Print a secret provided by the host - `curl -X GET -i http://127.0.0.1:8000/http-python/secret`
 
 # Extending RVM
 
@@ -61,3 +62,11 @@ After you are done with the host changes, you need to make the changes available
 ## Changing backing store
 We use OpenDAL, so switching backing store to something that's not the file system only requires you to use another service.
 
+## Cheat sheet (debug build)
+
+```
+cargo r
+cargo b -p http-rust --target wasm32-wasip2
+curl --data-binary "@target/wasm32-wasip2/debug/http_rust.wasm" localhost:8002/deploy/http-rust
+curl -X POST -i http://127.0.0.1:8000/http-rust/echo -d "echo this!"
+```
