+++ guests/http-rust/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "http-rust"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+wit-bindgen = "0.47.0"
+futures = "0.3"
