--- guests/rust/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "rust-svc"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-# [lib]
-# crate-type = ["cdylib"]
-
-[package.metadata.component]
-package = "component:rust-svc"
-# This tells cargo-component to use a special adapter, which doesn't depend on
-# `get-environment` or other things not present in the proxy world.
-proxy = true
-
-[dependencies]
-wit-bindgen-rt = { version = "0.41.0", features = ["bitflags"] }
-wstd = "0.5.4"
-
-[build]
-target = "wasm32-wasip2"
