--- guests/rust/README.md
@@ -1,7 +1,0 @@
-## Build
-
-`cargo build --target wasm32-wasip2`
-
-## 
-`guests/rust/target/wasm32-wasip2/debug/base-svc.wasm`
-
