+++ src/lib.rs
@@ -1,0 +1,14 @@
+use wasmtime::component::bindgen;
+
+pub mod host;
+pub mod state;
+
+bindgen!({
+    path: "./wit",
+    world: "rvm",
+    async: true,
+    with: {
+        "wasi:http/types@0.2.3": wasmtime_wasi_http::bindings::http::types,
+        "wasi:http@0.2.3": wasmtime_wasi_http::bindings::http,
+    }
+});
